requirements.txt
Flask==2.0.3
grpcio==1.44.0
grpcio-tools==1.44.0
